move train_image_file_path from models
